withoutleaks1

+++ LibraryC/array.c
@@ -23,7 +23,7 @@ Array* array_create(size_t size, FFree f)
 // delete array, free memory
 void array_delete(Array* arr)
 {
-    for (int i = 0; i < arr->size;i++) {
+    for (size_t i = 0; i < arr->size;i++) {
         free(arr->array[i]);
         arr->array[i] = NULL;
     }
